source/requirements.txt
Django==1.11.5
djangorestframework==3.6.4
requests==2.18.4
